projects/01-azure-openai-rag/requirements.txt
fastapi
uvicorn
openai
azure-search-documents
azure-core
pydantic
python-dotenv
httpx
